Rename windows module to backend_sdl
Try to have all the sdl2 code in the same module
