+++ src/backend_sdl/slideshow.rs
