+++ src/main.rs
@@ -24,8 +24,8 @@ use structopt::StructOpt;
 
 // Slidy imports.
 
-use slidy::windows::slideshow::SlideShowWindow;
-use slidy::windows::timer::TimerWindow;
+use slidy::backend_sdl::slideshow::SlideShowWindow;
+use slidy::backend_sdl::timer::TimerWindow;
 
 // Local modules.
 
@@ -161,9 +161,9 @@ fn main() {
     // since I want the slider to live on another thread.
 
     // Init stuffs
-    let sdl_context = slidy::get_sdl_context();
-    let ttf_context = slidy::get_ttf_context();
-    let free_mono = slidy::get_default_font(&ttf_context);
+    let sdl_context = slidy::backend_sdl::get_sdl_context();
+    let ttf_context = slidy::backend_sdl::get_ttf_context();
+    let free_mono = slidy::backend_sdl::get_default_font(&ttf_context);
 
     // 1. The slideshow window
     let mut slideshow_win = SlideShowWindow::new(
