--- src/windows/mod.rs
@@ -1,3 +1,0 @@
-pub mod slideshow;
-pub mod timer;
-pub mod utils;
